# Documentazione/apidocs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /me/profile:
    get:
      tags:
      - Profile
      summary: Leggi il tuo profilo
      description: Restituisce i dati anagrafici dell'utente autenticato.
      operationId: getMyProfile
      responses:
        "200":
          description: Profilo trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
        "404":
          description: Profilo non ancora creato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
        "401":
          description: Non autenticato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
    put:
      tags:
      - Profile
      summary: Aggiorna il tuo profilo
      description: Modifica i dati anagrafici dell'utente autenticato.
      operationId: updateMyProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileRequestDTO"
        required: true
      responses:
        "200":
          description: Profilo aggiornato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
        "400":
          description: Dati non validi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
        "404":
          description: Profilo non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
        "401":
          description: Non autenticato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
    post:
      tags:
      - Profile
      summary: Crea il tuo profilo
      description: Crea il profilo anagrafico per l'utente autenticato. Può essere
        fatto una sola volta.
      operationId: createMyProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileRequestDTO"
        required: true
      responses:
        "201":
          description: Profilo creato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
        "400":
          description: Dati non validi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
        "409":
          description: Profilo già esistente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
        "401":
          description: Non autenticato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponseDTO"
...
  /admin/venues/{id}:
    put:
      tags:
      - Venues
      summary: Modifica una sede esistente
      description: Solo ADMIN.
      operationId: updateVenue
      parameters:
      - name: id
        in: path
        description: ID sede
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueRequestDTO"
        required: true
      responses:
        "200":
          description: Sede aggiornata
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponseDTO"
        "400":
          description: Dati non validi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponseDTO"
        "404":
          description: Sede non trovata
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponseDTO"
    delete:
      tags:
      - Venues
      summary: Elimina una sede
      description: Solo ADMIN.
      operationId: deleteVenue
      parameters:
      - name: id
        in: path
        description: ID sede
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "204":
          description: Sede eliminata
        "404":
          description: Sede non trovata
  /admin/users/{id}/reactivate:
    put:
      tags:
      - Users
      summary: Riattiva un utente bannato
      description: Solo ADMIN.
      operationId: reactivateUser
      parameters:
      - name: id
        in: path
        description: ID utente da riattivare
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Utente riattivato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: Utente non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
  /admin/users/{id}/promote:
    put:
      tags:
      - Users
      summary: Promuove un utente a ORGANIZER
      description: Solo ADMIN. Applicabile solo a utenti con ruolo USER.
      operationId: promoteUser
      parameters:
      - name: id
        in: path
        description: ID utente da promuovere
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Ruolo aggiornato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: Utente non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "409":
          description: Utente non è USER o è ADMIN
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
  /admin/users/{id}/ban:
    put:
      tags:
      - Users
      summary: Banna un utente
      description: Solo ADMIN. Non applicabile agli ADMIN. Cancella automaticamente
        i biglietti futuri dell'utente.
      operationId: banUser
      parameters:
      - name: id
        in: path
        description: ID utente da bannare
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Utente bannato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: Utente non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "409":
          description: Operazione non consentita (es. target è ADMIN)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
...
  /admin/venues:
    post:
      tags:
      - Venues
      summary: Crea una nuova sede
      description: Solo ADMIN.
      operationId: createVenue
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueRequestDTO"
        required: true
      responses:
        "201":
          description: Sede creata
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponseDTO"
        "400":
          description: Dati non validi
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponseDTO"
        "409":
          description: Sede con lo stesso nome già esistente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponseDTO"
...
  /venues/{id}:
    get:
      tags:
      - Venues
      summary: Dettaglio sede per ID
      operationId: getVenueById
      parameters:
      - name: id
        in: path
        description: ID sede
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Sede trovata
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponseDTO"
        "404":
          description: Sede non trovata
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VenueResponseDTO"
...
  /me:
    get:
      tags:
      - Users
      summary: Dati utente corrente
      description: "Restituisce email, ruolo e status dell'utente autenticato."
      operationId: getMe
      responses:
        "200":
          description: Dati restituiti
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "401":
          description: Non autenticato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
...
  /admin/users:
    get:
      tags:
      - Users
      summary: Lista tutti gli utenti
      description: Solo ADMIN.
      operationId: getAllUsers
      responses:
        "200":
          description: Lista utenti
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponseDTO"
        "403":
          description: Accesso negato
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponseDTO"
  /admin/users/{id}:
    get:
      tags:
      - Users
      summary: Dettaglio utente per ID
      description: Solo ADMIN.
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: ID utente
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Utente trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "404":
          description: Utente non trovato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "403":
          description: Accesso negato
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
...
components:
  schemas:
    UserProfileRequestDTO:
      type: object
      description: Dati per la creazione o aggiornamento del profilo utente
      properties:
        firstName:
          type: string
          description: Nome
          example: Mario
          minLength: 1
        lastName:
          type: string
          description: Cognome
          example: Rossi
          minLength: 1
        bio:
          type: string
          description: Biografia
          example: Appassionato di eventi tech
        city:
          type: string
          description: Città
          example: Roma
        photoUrl:
          type: string
          description: URL foto profilo
          example: https://example.com/foto.jpg
      required:
      - firstName
      - lastName
    UserProfileResponseDTO:
      type: object
      description: Dati del profilo utente
      properties:
        id:
          type: integer
          format: int32
          description: ID profilo
          example: 1
        firstName:
          type: string
          description: Nome
          example: Mario
        lastName:
          type: string
          description: Cognome
          example: Rossi
        bio:
          type: string
          description: Biografia
          example: Appassionato di eventi tech
        city:
          type: string
          description: Città
          example: Roma
        photoUrl:
          type: string
          description: URL foto profilo
          example: https://example.com/foto.jpg
...
    UserResponseDTO:
      type: object
      description: Dati pubblici di un utente
      properties:
        id:
          type: integer
          format: int32
          description: ID utente
          example: 1
        email:
          type: string
          description: Email utente
          example: mario@example.com
        role:
          type: string
          description: Ruolo utente
          enum:
          - USER
          - ORGANIZER
          - ADMIN
          example: USER
        status:
          type: string
          description: Stato utente
          enum:
          - ACTIVE
          - BANNED
          example: ACTIVE
    VenueResponseDTO:
      type: object
      description: Dati di una sede
      properties:
        id:
          type: integer
          format: int32
          description: ID sede
          example: 1
        name:
          type: string
          description: Nome della sede
          example: Centro Congressi Roma
        address:
          type: string
          description: Indirizzo
          example: "Via Roma 1, Roma"
        capacity:
          type: integer
          format: int32
          description: Capienza massima
          example: 200
    VenueRequestDTO:
      type: object
      description: Dati per la creazione o modifica di una sede
      properties:
        name:
          type: string
          description: Nome della sede
          example: Centro Congressi Roma
          minLength: 1
        address:
          type: string
          description: Indirizzo
          example: "Via Roma 1, Roma"
          minLength: 1
        capacity:
          type: integer
          format: int32
          description: Capienza massima
          example: 200
          minimum: 1
      required:
      - address
      - capacity
      - name
...